move spec editor sidebar
