+++ packages/insomnia/src/ui/components/spec-editor/sidebar/sidebar-header.tsx
@@ -2,7 +2,7 @@ import { motion } from 'framer-motion';
 import React, { FunctionComponent, ReactNode } from 'react';
 import styled from 'styled-components';
 
-import { IconEnum, SvgIcon } from '../svg-icon';
+import { IconEnum, SvgIcon } from '../../../insomnia-components/svg-icon';
 
 export interface SidebarHeaderProps {
   headerTitle: string;
