+++ packages/insomnia/src/ui/components/spec-editor/sidebar/sidebar-item.tsx
