+++ packages/insomnia/src/ui/components/spec-editor/sidebar/sidebar-parameters.tsx
@@ -1,7 +1,7 @@
 import React, { Component, Fragment, ReactNode } from 'react';
 
-import { Tooltip } from '../../components/tooltip';
-import { IconEnum, SvgIcon } from '../svg-icon';
+import { Tooltip } from '../../tooltip';
+import { IconEnum, SvgIcon } from '../../../insomnia-components/svg-icon';
 import { SidebarInvalidSection } from './sidebar-invalid-section';
 import { SidebarItem } from './sidebar-item';
 import { SidebarSection } from './sidebar-section';
