+++ packages/insomnia/src/ui/components/spec-editor/sidebar/sidebar-text-item.tsx
